# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/task1/public/orders?page={page}&limit={limit}": {
            "get": {
                "tags": [
                    "List Orders"
                ],
                "summary": "Get list of orders",
                "description": "Returns list of orders",
                "operationId": "getOrdersList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit Number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/task1/public/order": {
            "post": {
                "tags": [
                    "Create Order"
                ],
                "summary": "Add a new order to the store",
                "description": "",
                "operationId": "addOrder",
                "requestBody": {
                    "description": "Order object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin": {
                                        "description": "Origin Coordinates",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "destination": {
                                        "description": "Destination Coordinates",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:orders",
                            "read:orders"
                        ]
                    }
                ]
            }
        },
        "/task1/public/order/{orderId}": {
            "put": {
                "tags": [
                    "Order Update"
                ],
                "summary": "Update an existing order",
                "description": "",
                "operationId": "updatePet",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "Order id to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Order object to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "Status",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "orderstore_auth": [
                            "write:orders",
                            "read:orders"
                        ]
                    }
                ]
            }
        }
    }
}